Remove Blaze repo based dependencies from static-html
The `static-html` package really only needs the following Blaze
repo based functionality to accomplish its role:

1. [`caching-html-compiler`](https://github.com/meteor/blaze/tree/555f3bb0d34c7553423e1f01f054189378e343ea/packages/caching-html-compiler)
2. The [`html-scanner.js`](https://github.com/meteor/blaze/blob/555f3bb0d34c7553423e1f01f054189378e343ea/packages/templating-tools/html-scanner.js) file from the [`templating-tools`](https://github.com/meteor/blaze/tree/555f3bb0d34c7553423e1f01f054189378e343ea/packages/templating-tools) package.
3. And of course itself, the [`static-html`](https://github.com/meteor/blaze/tree/555f3bb0d34c7553423e1f01f054189378e343ea/packages/static-html) package.

Since the above functionality isn't specific to Blaze, this
commit creates a new copy of the `static-html` and
`caching-html-compiler` packages in core, and creates a new
package called `html-scanner`, to house the `html-scanner.js`
functionality from the `templating-tools` package. With these
changes in-place, the `static-html` package can be used
without pulling in a bunch of un-necessary Blaze based
utilities / functionality.