Bump $BUNDLE_VERSION to 8.11.1.1 before rebuilding dev bundle.
The dev bundle needs to be rebuilt after 1.6.1 since we bumped the Node
version in 1.6.1 but did not bump the npm version, though npm has been
updated to 5.8.0 on the devel branch (and release-1.6.2) already.

The $BUNDLE_VERSION version is interpreted as an opaque string rather than
a semantic version, so we can easily add a fourth number to track changes
that fall between Node versions (while we're waiting for Node 8.11.2).