Document `renderToNodeStream`
I switched `renderToString` to `renderToNodeStream` and noticed that the stylesheet that styled-components generate is empty.

This happens when using `sink.appendToHead(sheet.getStyleTags());` alongside `renderToNodeStream`.

The Solution is to use the new `sheet.interleaveWithNodeStream`. With this approach, the styles are appended to the body instead of the head, but this should have no impact.

See also https://www.styled-components.com/docs/advanced#streaming-rendering

Would be good to have a meteor-section on that page as well ;-) Maybe someone can open a PR there?