Prefer client.s.databaseName over client.s.options.dbName.
Fixes #9827.

The client.s.databaseName property appears to have been introduced a long
time ago (in 2014), so it seems reliable:
https://github.com/mongodb/node-mongodb-native/commit/14fd60b99d92760e130e67e187b4c22dfd74ea7d

The use of client.s.options.dbName was only recently introduced in #9790,
since the MongoDB.connect callback now receives a MongoClient object
rather than a Db object. Not sure if client.s.options.dbName was ever
reliable, but at least we know when the problem started.