tests: Add additional self-tests to the AppVeyor suite which exercise Mongo.
Our current testing on Mongo is less complete than it could be and these
tests, particularly the `mongo with multiple --port numbers` test which does
actual writes to the database, should help with that.

Note that these tests are separate from the `mongo` package tests which are
much more extensive and test more advanced mongo functionality, but it's
good to have some basic Windows tests in place within the `self-test`
infrastructure.

cc @benjamn