Rename options.cacheDir to deps.cacheDirectory.
The options passed to meteorBabel.compile are passed along to Babel, and
cacheDir is not a valid Babel compilation option.

The deps argument is used to compute cache keys, so including the cache
directory path seems appropriate.

This API could be simplified if we were willing to break backwards
compatibility, but we're not there yet.