Tolerate multiple meteorInstall calls when scanning bundle stats.
Because of 8b04c253900e4ca2a194d2fcaf6fc2ce9a9085e7, there may now be
multiple calls to meteorInstall in the modules package bundle.

The code for extracting module sizes from the minified bundle previously
assumed there would be only one call to meteorInstall, so this commit
allows for more than one call.

While I was at it, I deleted the simplistic Visitor implementation this
code was using in favor of the more sophisticated Visitor provided by the
reify package, which has the abitlity to skip whole subtrees that do not
contain the identifier(s) we care about.