Use Npm.depends to install core-js in ecmascript-runtime-client.
The @babel/plugin-transform-runtime@7.0.0-beta.56 plugin used by
meteor-babel@7.0.0-beta.56 now generates code that imports helpers from
@babel/runtime@7.0.0-beta.56, which no longer relies on core-js:
https://github.com/babel/babel/pull/8266
https://github.com/babel/babel/releases/tag/v7.0.0-beta.56

This is not much of a change for modern browsers, since we were already
using @babel/runtime/helpers/builtin to avoid relying on core-js. However,
for the legacy bundle, it's now more important than ever that we import
appropriate core-js polyfills in ecmascript-runtime-client, since they are
not going to be imported as a side-effect of using @babel/runtime.

That means there's no longer as much benefit to sharing the application's
copy of core-js with ecmascript-runtime-client, so it seems simpler just
to give ecmascript-runtime-client its own private copy of core-js, so that
it can import appropriate polyfills without forcing the developer to
install core-js explicitly in the application.