Prevent from using incompletely downloaded files (#38)
* Prevent from using incompletely downloaded files

This deals with the situation in which an incompletely downloaded file
is saved to disk and later treated like a fully downloaded one. This
can happen on slow, disconnecting internet connections.

* Prevent from using incompletely downloaded files

This deals with the situation in which an incompletely downloaded file is saved to disk and later treated like a fully downloaded one. This can happen on slow, disconnecting internet connections.